Rename utils/log_actions.py to utils/logging.py
