+++ wagtail/actions/copy_page.py
@@ -7,7 +7,7 @@ from modelcluster.models import get_all_child_relations
 from wagtail.models.i18n import TranslatableMixin
 from wagtail.signals import page_published
 from wagtail.utils.copying import _copy, _copy_m2m_relations
-from wagtail.utils.log_actions import log
+from wagtail.utils.logging import log
 
 logger = logging.getLogger("wagtail")
