+++ wagtail/actions/create_alias.py
@@ -5,7 +5,7 @@ from django.core.exceptions import PermissionDenied
 
 from wagtail.models.i18n import TranslatableMixin
 from wagtail.utils.copying import _copy, _copy_m2m_relations
-from wagtail.utils.log_actions import log
+from wagtail.utils.logging import log
 
 logger = logging.getLogger("wagtail")
