+++ wagtail/actions/move_page.py
@@ -5,7 +5,7 @@ from django.db import transaction
 from treebeard.mp_tree import MP_MoveHandler
 
 from wagtail.signals import post_page_move, pre_page_move
-from wagtail.utils.log_actions import log
+from wagtail.utils.logging import log
 
 logger = logging.getLogger("wagtail")
