+++ wagtail/admin/auth.py
@@ -13,7 +13,7 @@ from django.utils.translation import override
 
 from wagtail.admin import messages
 from wagtail.models import GroupPagePermission
-from wagtail.utils.log_actions import LogContext
+from wagtail.utils.logging import LogContext
 
 
 def users_with_page_permission(page, permission_type, include_superusers=True):
