+++ wagtail/admin/views/generic/models.py
@@ -14,7 +14,7 @@ from django.views.generic.list import BaseListView
 
 from wagtail.admin import messages
 from wagtail.admin.ui.tables import Table, TitleColumn
-from wagtail.utils.log_actions import log
+from wagtail.utils.logging import log
 
 from .base import WagtailAdminTemplateMixin
 from .permissions import PermissionCheckedMixin
