+++ wagtail/contrib/modeladmin/views.py
@@ -41,8 +41,8 @@ from wagtail.admin import messages
 from wagtail.admin.ui.tables import Column, DateColumn, Table, UserColumn
 from wagtail.admin.views.generic.base import WagtailAdminTemplateMixin
 from wagtail.admin.views.mixins import SpreadsheetExportMixin
-from wagtail.utils.log_actions import log
-from wagtail.utils.log_actions import registry as log_registry
+from wagtail.utils.logging import log
+from wagtail.utils.logging import registry as log_registry
 
 from .forms import ParentChooserForm
